# authcenter/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "通用認證中心",
        "title": "authcenter",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/create-vertify-email": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Register"
                ],
                "summary": "創建email認證連結",
                "parameters": [
                    {
                        "description": "email",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_RoyceAzure_lab_authcenter_internal_api_dto.CreateVertifyUserEmailLinkDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "UnauthenticatedCode",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.ResponseError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "460": {
                        "description": "InvalidArgumentCode",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.ResponseError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.ResponseError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/linkedUser": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Register"
                ],
                "summary": "使用帳號密碼連結email，必須要先認證email",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "accountInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_RoyceAzure_lab_authcenter_internal_api_dto.LinkedUserAccountAndPasswordDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.ResponseError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/login/account": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "account and password login",
                "parameters": [
                    {
                        "description": "account and password",
                        "name": "accountInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_RoyceAzure_lab_authcenter_internal_api_dto.AccountAndPasswordLoginDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_RoyceAzure_lab_authcenter_internal_api_dto.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "UnauthenticatedCode",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.ResponseError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "UnauthorizedCode",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.ResponseError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.ResponseError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/login/google": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "google login",
                "parameters": [
                    {
                        "description": "google id token",
                        "name": "id_token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_RoyceAzure_lab_authcenter_internal_api_dto.GoogleLoginDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_RoyceAzure_lab_authcenter_internal_api_dto.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "UnauthenticatedCode",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.ResponseError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "UnauthorizedCode",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.ResponseError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.ResponseError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "logout",
                "parameters": [
                    {
                        "description": "refresh token",
                        "name": "refresh_token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_RoyceAzure_lab_authcenter_internal_api_dto.RefreshTokenDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "UnauthenticatedCode",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.ResponseError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "UnauthorizedCode",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.ResponseError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.ResponseError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.MetaData": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "api.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "不要使用omitempty 否則空list也會回傳null"
                },
                "meta": {
                    "$ref": "#/definitions/api.MetaData"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_RoyceAzure_lab_authcenter_internal_api_dto.AccountAndPasswordLoginDTO": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "帳號",
                    "type": "string"
                },
                "password": {
                    "description": "密碼明文",
                    "type": "string"
                }
            }
        },
        "github_com_RoyceAzure_lab_authcenter_internal_api_dto.CreateVertifyUserEmailLinkDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "github_com_RoyceAzure_lab_authcenter_internal_api_dto.GoogleLoginDTO": {
            "type": "object",
            "properties": {
                "id_token": {
                    "type": "string"
                }
            }
        },
        "github_com_RoyceAzure_lab_authcenter_internal_api_dto.LinkedUserAccountAndPasswordDTO": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_RoyceAzure_lab_authcenter_internal_api_dto.LoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "$ref": "#/definitions/github_com_RoyceAzure_lab_authcenter_internal_api_dto.TokenInfo"
                },
                "refresh_token": {
                    "$ref": "#/definitions/github_com_RoyceAzure_lab_authcenter_internal_api_dto.TokenInfo"
                },
                "user": {
                    "$ref": "#/definitions/github_com_RoyceAzure_lab_authcenter_internal_api_dto.UserDTO"
                }
            }
        },
...
        "github_com_RoyceAzure_lab_authcenter_internal_api_dto.RefreshTokenDTO": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "github_com_RoyceAzure_lab_authcenter_internal_api_dto.TokenInfo": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Description for Authorization header: Type \"Bearer\" followed by a space and the token. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}